Fix: 삭제 및 탈퇴시 alert창 띄우기

+++ public/JS/detailpage.js
@@ -11,7 +11,7 @@ const clip = () => {
 };
 
 // board 정보 가져오기
-const mainElement = document.querySelector('main');
+const mainElement = document.querySelector("main");
 const boardId = mainElement.dataset.boardId;
 let likeCount = parseInt(mainElement.dataset.likeCount);
 let likedByUser = mainElement.dataset.likedByUser;
@@ -36,7 +36,6 @@ const countFavorite = async (boardId) => {
         method: "delete",
         url: `/like/delete/${boardId}`,
       }).then((res) => {
-        alert("삭제완료");
         likeCount--;
         likedByUser = false;
       });
@@ -47,7 +46,6 @@ const countFavorite = async (boardId) => {
         url: "/like/post",
         data: { boardId },
       }).then((res) => {
-        alert("추가완료");
         likeCount++;
         likedByUser = true;
       });
@@ -61,27 +59,49 @@ const countFavorite = async (boardId) => {
 
 const updateUI = (likeCount, likedByUser) => {
   document.querySelector(".likecount").innerHTML = `좋아요수 ${likeCount}`;
-  document.querySelector("heart").src = likedByUser
+  document.querySelector(".heart").src = likedByUser
     ? "/public/image/fillheart.svg"
     : "/public/image/binheart.svg";
 };
 
 // 글 삭제
 const deleteBoard = (boardId) => {
-  axios({
-    method: "delete",
-    url: `/board/delete/${boardId}`,
-  })
-    .then((res) => {
-      if (res.data.result) {
-        window.location.href = "http://localhost:3000/";
-      } else {
-        console.log("실패");
-      }
-    })
-    .catch((e) => {
-      console.log("삭제 실패", e);
-    });
+  Swal.fire({
+    title: "게시물 삭제",
+    text: "진짜로 삭제할건가요??",
+    icon: "question",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "네",
+    cancelButtonText: "아니요",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      axios({
+        method: "delete",
+        url: `/board/delete/${boardId}`,
+      })
+        .then((res) => {
+          if (res.data.result) {
+            Swal.fire({
+              title: "삭제 완료",
+              text: "삭제되었습니다!",
+              icon: "success",
+              confirmButtonText: "확인",
+            }).then((result) => {
+              if (result.isConfirmed) {
+                window.location.href = "http://localhost:3000/";
+              }
+            });
+          } else {
+            console.log("실패");
+          }
+        })
+        .catch((e) => {
+          console.log("삭제 실패", e);
+        });
+    }
+  });
 };
 
 // 현재 좋아요 상태 호출
